pyproject.toml
[project]
name = "agentforge"
version = "0.1.0a1"
description = "Complete AI agent infrastructure in one command"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Jakebot Labs", email = "jakebot2ops@gmail.com"}]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "httpx>=0.26",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio"]

[project.scripts]
agentforge = "agentforge.cli:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
